Use new XmlLocator API to enable tooltips-on-hover for property text.
